Restart systemd-timedated after rw-remounting the root filesystem
The systemd service timedated is responsible for applying the time zone
settings. It is smart enough to detect the read-only filesystem upfront,
thus it needs to be restarted after the rw remount to be fully aware of
the fact it can write to the file system.

This commit also improves the clean up in the finally stanza to avoid
masking errors from the test if the revertion of the time zone to UTC
fails.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>