MEN-4769: include the current JWT token in the reported config checksum
Including the current JWT token in the reported config checksum makes
sure the mender-configure addon reports the configuration to the server
in case of a new authorization. This covers a corner case: without this
change, if the device gets decommissioned in the server, then accepted
again, the server will assign a new device ID to it, but no
configuration will be reported until new values are deployed to it.

Changelog: title

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>