Fix: mender-configure requests JWT token when needed
If the token is invalid or inexistant, mender-configure will fetch a new Jwt Token.
Mender configure will listen to the dbus-monitor for a Jwt token state change. If token is received, it will immediately get the new token and continue the execution of the program. If the token is not received within 5 seconds, the program times out.

Ticket: MEN-7732
Signed-off-by: Victor Moene <victor.moene@northern.tech>
Changelog: mender-configure doesn't rely anymore on mender-updated to fetch new Jwt tokens over dbus